Did some changes to facade design pattern

+++ DP/src/main/java/com/hotel/strategy/CreditCardStrategy.java
@@ -15,7 +15,7 @@ public class CreditCardStrategy implements PaymentStrategy{
     }
 
     @Override
-    public void pay(int amount) {
+    public void pay(long amount) {
         System.out.println(amount + " paid with credit/debit card");
     }
 }
